third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-bad-array-argument-expected.txt

Validates that lockOrientation() returns false on bad orientation value in array argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation is "portrait-primary"
PASS screen.lockOrientation(['portrait-primary', 'invalid-orientation']) is false
PASS screen.lockOrientation(['portrait-primary', null]) is false
PASS screen.lockOrientation(['portrait-primary', undefined]) is false
PASS screen.lockOrientation(['portrait-primary', 123]) is false
PASS screen.lockOrientation(['portrait-primary', window]) is false
PASS screen.orientation is "portrait-primary"
PASS successfullyParsed is true

TEST COMPLETE
